adjust TPC-H test
