+++ .github/workflows/tpc-h.yml
@@ -89,80 +89,34 @@ jobs:
               with:
                   DB-name: ${{ matrix.DB-name }}
 
-            - name: Get Old Version IGinX and Copy TPC-H Data
-              shell: bash
-              run: |
-                  git clone https://github.com/IGinX-THU/IGinX.git
-                  mkdir IGinX/tpc
-                  cp -r tpc/* IGinX/tpc/
-                  ls -a 'IGinX/tpc/TPC-H V3.0.1/data'
-
-            - name: Rewrite DB Conf in Old IGinX
-              uses: ./.github/actions/dbConfWriter
-              with:
-                  DB-name: ${{ matrix.DB-name }}
-                  Root-Dir-Path: "IGinX"
-
-            - name: Install Old IGinX with Maven
-              shell: bash
-              run: |
-                  cd IGinX
-                  mvn clean package -DskipTests -P-format -q
-
-            - name: Change Old IGinX config
-              uses: ./.github/actions/confWriter
-              with:
-                  DB-name: ${{ matrix.DB-name }}
-                  Set-Filter-Fragment-OFF: "true"
-                  Metadata: ${{ matrix.metadata }}
-                  Root-Dir-Path: "IGinX"
-
-            - name: Start Old IGinX
-              if: always()
-              shell: bash
-              run: |
-                  cd IGinX/core/target/iginx-core-${VERSION}
-                  pwd
-                  export IGINX_HOME=$PWD
-                  echo "IGinX home path: $IGINX_HOME"
-                  cd ..
-                  chmod +x iginx-core-${VERSION}/sbin/start_iginx.sh
-                  nohup iginx-core-${VERSION}/sbin/start_iginx.sh > ../../iginx-${VERSION}.log 2>&1 &
-
-            - name: Run Regression Test on Old IGinX
-              if: always()
-              shell: bash
-              run: |
-                  cd IGinX
-                  mvn test -q -Dtest=TPCHRegressionIT -DfailIfNoTests=false -P-format
-
-            - name: Show Old IGinX log
-              if: always()
-              shell: bash
-              run: |
-                  cat IGinX/iginx-*.log
-
-            - name: Stop ZooKeeper
-              uses: ./.github/actions/zookeeperRunner
-              with:
-                  if-stop: "true"
-
-            - name: Stop Old IGinX
-              uses: ./.github/actions/iginxRunner
-              with:
-                  version: ${VERSION}
-                  if-stop: "true"
-
-            - name: Rerun ZooKeeper
-              uses: ./.github/actions/zookeeperRunner
-              with:
-                  if-rerun: "true"
+            #            - name: Get Old Version IGinX
+            #              shell: bash
+            #              run: git clone https://github.com/IGinX-THU/IGinX.git
+            #
+            #            - name: Rewrite DB Conf in Old IGinX
+            #              uses: ./.github/actions/dbConfWriter
+            #              with:
+            #                  DB-name: ${{ matrix.DB-name }}
+            #                  Root-Dir-Path: "IGinX"
+            #
+            #            - name: Install Old IGinX with Maven
+            #              shell: bash
+            #              run: |
+            #                  cd IGinX
+            #                  mvn clean package -DskipTests -P-format -q
+            #
+            #            - name: Change Old IGinX config
+            #              uses: ./.github/actions/confWriter
+            #              with:
+            #                  DB-name: ${{ matrix.DB-name }}
+            #                  Set-Filter-Fragment-OFF: "true"
+            #                  Metadata: ${{ matrix.metadata }}
+            #                  Root-Dir-Path: "IGinX"
 
             - name: Install New IGinX with Maven
               shell: bash
               run: |
                   mvn clean package -DskipTests -P-format -q
-                  cp IGinX/test/src/test/resources/tpch/oldTimeCosts.txt test/src/test/resources/tpch/
 
             - name: Change New IGinX config
               uses: ./.github/actions/confWriter
@@ -174,14 +128,47 @@ jobs:
             - name: Start New IGinX
               uses: ./.github/actions/iginxRunner
 
-            - name: Run Regression Test on New IGinX
-              if: always()
+            - name: Insert Data and Register UDF
               shell: bash
-              run: |
-                  mvn test -q -Dtest=TPCHRegressionIT -DfailIfNoTests=false -P-format
+              run: mvn test -q -Dtest=TPCHDataGeneratorIT -DfailIfNoTests=false -P-format
+
+            - name: Stop New IGinX
+              uses: ./.github/actions/iginxRunner
+              with:
+                  version: ${VERSION}
+                  if-stop: "true"
 
             - name: Show New IGinX log
               if: always()
               shell: bash
-              run: |
-                  cat iginx-*.log
+              run: cat iginx-*.log
+
+            - name: Run Regression Test the 1st time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 2nd time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 3rd time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 4th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 5th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 6th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 7th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 8th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 9th time
+              uses: ./.github/actions/tpchSingleTest
+
+            - name: Run Regression Test the 10th time
+              uses: ./.github/actions/tpchSingleTest
