+++ test/src/test/java/cn/edu/tsinghua/iginx/integration/tool/ConfLoader.java
@@ -126,6 +126,16 @@ public class ConfLoader {
     return getTestProperty(DBCE_TEST_WAY, DEFAULT_DBCE_TEST_WAY);
   }
 
+  public List<Integer> getQueryIds() {
+    String input = properties.getProperty("query_ids");
+    String[] split = input.split(",");
+    List<Integer> list = new ArrayList<>();
+    for (String s : split) {
+      list.add(Integer.parseInt(s));
+    }
+    return list;
+  }
+
   public int getMaxRepetitionsNum() {
     return Integer.parseInt(properties.getProperty("max_repetitions_num"));
   }
