+++ test/src/test/resources/testConfig.properties
@@ -73,5 +73,6 @@ is_scaling=false
 DBCE_test_way=oriHasDataExpHasData
 
 # TPC-H test
+query_ids=1,2,3,5,6,9,10,16,17,18,19,20
 max_repetitions_num=10
 regression_threshold=0.3
