+++ test/src/test/resources/tpch/runtimeInfo/failedQueryIds.txt
@@ -1,0 +1,12 @@
+1
+2
+3
+5
+6
+9
+10
+16
+17
+18
+19
+20
