+++ test/src/test/resources/tpch/runtimeInfo/iterationTimes.txt
@@ -1,0 +1,1 @@
+1
